fastapi_website/requirements.txt
# pip install -r requirements.txt

# FastAPI --------------------------------------
fastapi
uvicorn
jinja2
aiofiles
python-multipart
requests

# Other stuff from https://github.com/ultralytics/yolov5/blob/master/requirements.txt
# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=2.2
torch>=1.7.0
tqdm>=4.41.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP
torchvision>=0.8.1